ENG-01: extract _nipals; drop dead code flagged by CodeQL
Move the NIPALS / least-squares kernels (nan_to_zeros,
regress_a_space_on_b_row, ssq, terminate_check, quick_regress,
internal_pls_nipals_fit_one_pc) into _nipals, and discard the long-dead
commented-out _apply_pca block. Also remove the commented-out 'options'
block in _preprocessing.scale that CodeQL flagged as commented-out code.
